Add logging, version info, and API cleanup
Introduces detailed logging for engine, SDL, SDL_image, SDL_ttf, audio, and GPU device information. Refactors the API to remove the 'debug' argument from window creation and instead adds a 'debug' flag to engine initialization. Changes animation FPS types from int to double for greater precision. Bumps version to 1.4.0 and updates SDL3 dependency to preview-3.3.4.

+++ include/AnimationController.hpp
@@ -27,14 +27,14 @@ struct Frame
 struct Animation
 {
     std::vector<Frame> frames;
-    int fps;
+    double fps;
 };
 
 struct SheetStrip
 {
     std::string name;
     int frameCount;
-    int fps;
+    double fps;
 };
 
 class AnimationController
